Fix: Migrate to authdata for better set/get key security, split up auth into files, added preliminary BackupAPI, use Result instead of Resource. (#1822)
