+++ app/src/main/java/com/lagradost/cloudstream3/syncproviders/AccountManager.kt
@@ -1,0 +1,158 @@
+package com.lagradost.cloudstream3.syncproviders
+
+import com.lagradost.cloudstream3.AcraApplication.Companion.getKey
+import com.lagradost.cloudstream3.AcraApplication.Companion.setKey
+import com.lagradost.cloudstream3.LoadResponse
+import com.lagradost.cloudstream3.syncproviders.providers.Addic7ed
+import com.lagradost.cloudstream3.syncproviders.providers.AniListApi
+import com.lagradost.cloudstream3.syncproviders.providers.LocalList
+import com.lagradost.cloudstream3.syncproviders.providers.MALApi
+import com.lagradost.cloudstream3.syncproviders.providers.OpenSubtitlesApi
+import com.lagradost.cloudstream3.syncproviders.providers.SimklApi
+import com.lagradost.cloudstream3.syncproviders.providers.SubDlApi
+import com.lagradost.cloudstream3.syncproviders.providers.SubSourceApi
+import com.lagradost.cloudstream3.utils.DataStoreHelper
+import java.util.concurrent.TimeUnit
+
+abstract class AccountManager {
+    companion object {
+        const val NONE_ID: Int = -1
+        val malApi = MALApi()
+        val aniListApi = AniListApi()
+        val simklApi = SimklApi()
+        val localListApi = LocalList()
+
+        val openSubtitlesApi = OpenSubtitlesApi()
+        val addic7ed = Addic7ed()
+        val subDlApi = SubDlApi()
+        val subSourceApi = SubSourceApi()
+
+        var cachedAccounts: MutableMap<String, Array<AuthData>>
+        var cachedAccountIds: MutableMap<String, Int>
+
+        const val ACCOUNT_TOKEN = "auth_tokens"
+        const val ACCOUNT_IDS = "auth_ids"
+
+        fun accounts(prefix: String): Array<AuthData> {
+            require(prefix != "NONE")
+            return getKey<Array<AuthData>>(
+                ACCOUNT_TOKEN,
+                "${prefix}/${DataStoreHelper.currentAccount}"
+            ) ?: arrayOf()
+        }
+
+        fun updateAccounts(prefix: String, array: Array<AuthData>) {
+            require(prefix != "NONE")
+            setKey(ACCOUNT_TOKEN, "${prefix}/${DataStoreHelper.currentAccount}", array)
+            synchronized(cachedAccounts) {
+                cachedAccounts[prefix] = array
+            }
+        }
+
+        fun updateAccountsId(prefix: String, id: Int) {
+            require(prefix != "NONE")
+            setKey(ACCOUNT_IDS, "${prefix}/${DataStoreHelper.currentAccount}", id)
+            synchronized(cachedAccountIds) {
+                cachedAccountIds[prefix] = id
+            }
+        }
+
+        val allApis = arrayOf(
+            SyncRepo(malApi),
+            SyncRepo(aniListApi),
+            SyncRepo(simklApi),
+            SyncRepo(localListApi),
+
+            SubtitleRepo(openSubtitlesApi),
+            SubtitleRepo(addic7ed),
+            SubtitleRepo(subDlApi),
+            SubtitleRepo(subSourceApi)
+        )
+
+        fun updateAccountIds() {
+            val ids = mutableMapOf<String, Int>()
+            for (api in allApis) {
+                ids.put(
+                    api.idPrefix,
+                    getKey<Int>(
+                        ACCOUNT_IDS,
+                        "${api.idPrefix}/${DataStoreHelper.currentAccount}",
+                        NONE_ID
+                    ) ?: NONE_ID
+                )
+            }
+            synchronized(cachedAccountIds) {
+                cachedAccountIds = ids
+            }
+        }
+
+        init {
+            val data = mutableMapOf<String, Array<AuthData>>()
+            val ids = mutableMapOf<String, Int>()
+            for (api in allApis) {
+                data.put(api.idPrefix, accounts(api.idPrefix))
+                ids.put(
+                    api.idPrefix,
+                    getKey<Int>(
+                        ACCOUNT_IDS,
+                        "${api.idPrefix}/${DataStoreHelper.currentAccount}",
+                        NONE_ID
+                    ) ?: NONE_ID
+                )
+            }
+            cachedAccounts = data
+            cachedAccountIds = ids
+        }
+
+        // I do not want to place this in the init block as JVM initialization order is weird, and it may cause exceptions
+        // accessing other classes
+        fun initMainAPI() {
+            LoadResponse.malIdPrefix = malApi.idPrefix
+            LoadResponse.aniListIdPrefix = aniListApi.idPrefix
+            LoadResponse.simklIdPrefix = simklApi.idPrefix
+        }
+
+        val subtitleProviders = arrayOf(
+            SubtitleRepo(openSubtitlesApi),
+            SubtitleRepo(addic7ed),
+            SubtitleRepo(subDlApi),
+            SubtitleRepo(subSourceApi)
+        )
+        val syncApis = arrayOf(
+            SyncRepo(malApi),
+            SyncRepo(aniListApi),
+            SyncRepo(simklApi),
+            SyncRepo(localListApi)
+        )
+
+        const val APP_STRING = "cloudstreamapp"
+        const val APP_STRING_REPO = "cloudstreamrepo"
+        const val APP_STRING_PLAYER = "cloudstreamplayer"
+
+        // Instantly start the search given a query
+        const val APP_STRING_SEARCH = "cloudstreamsearch"
+
+        // Instantly resume watching a show
+        const val APP_STRING_RESUME_WATCHING = "cloudstreamcontinuewatching"
+
+        fun secondsToReadable(seconds: Int, completedValue: String): String {
+            var secondsLong = seconds.toLong()
+            val days = TimeUnit.SECONDS
+                .toDays(secondsLong)
+            secondsLong -= TimeUnit.DAYS.toSeconds(days)
+
+            val hours = TimeUnit.SECONDS
+                .toHours(secondsLong)
+            secondsLong -= TimeUnit.HOURS.toSeconds(hours)
+
+            val minutes = TimeUnit.SECONDS
+                .toMinutes(secondsLong)
+            secondsLong -= TimeUnit.MINUTES.toSeconds(minutes)
+            if (minutes < 0) {
+                return completedValue
+            }
+            //println("$days $hours $minutes")
+            return "${if (days != 0L) "$days" + "d " else ""}${if (hours != 0L) "$hours" + "h " else ""}${minutes}m"
+        }
+    }
+}
