+++ app/src/main/java/com/lagradost/cloudstream3/syncproviders/SyncRepo.kt
@@ -1,0 +1,30 @@
+package com.lagradost.cloudstream3.syncproviders
+
+/** Stateless safe abstraction of SyncAPI */
+class SyncRepo(override val api: SyncAPI) : AuthRepo(api) {
+    val syncIdName = api.syncIdName
+    var requireLibraryRefresh: Boolean
+        get() = api.requireLibraryRefresh
+        set(value) {
+            api.requireLibraryRefresh = value
+        }
+
+    suspend fun updateStatus(id: String, newStatus: SyncAPI.AbstractSyncStatus): Result<Boolean> =
+        runCatching {
+            val status = api.updateStatus(freshAuth() ?: return@runCatching false, id, newStatus)
+            requireLibraryRefresh = true
+            status
+        }
+
+    suspend fun status(id: String): Result<SyncAPI.AbstractSyncStatus?> = runCatching {
+        api.status(freshAuth(), id)
+    }
+
+    suspend fun load(id: String): Result<SyncAPI.SyncResult?> = runCatching {
+        api.load(freshAuth(), id)
+    }
+
+    suspend fun library(): Result<SyncAPI.LibraryMetadata?> = runCatching {
+        api.library(freshAuth())
+    }
+}
